output consistency: refactoring: split selection file
