+++ misc/python/materialize/output_consistency/execution/query_execution_manager.py
@@ -27,7 +27,7 @@ from materialize.output_consistency.query.query_result import (
     QueryResult,
 )
 from materialize.output_consistency.query.query_template import QueryTemplate
-from materialize.output_consistency.selection.selection import (
+from materialize.output_consistency.selection.column_selection import (
     ALL_QUERY_COLUMNS_BY_INDEX_SELECTION,
 )
 from materialize.output_consistency.status.test_summary import ConsistencyTestSummary
